requirements.txt
# 龙头战法Web平台 + Skill 统一依赖

# Web后端
fastapi==0.133.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.3.4
python-multipart==0.0.22
python-dotenv==1.2.1

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
asyncpg==0.29.0

# 定时任务
apscheduler==3.10.4

# 数据源
adata>=1.0.0
tushare>=1.3.0

# DeepAgents（包含 LangChain）
deepagents==0.4.3
langchain-openai==0.1.0

# Skill脚本 + 通用
requests==2.32.5
pyyaml==6.0.3

# 测试
pytest==8.3.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
